% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeErrorProb.R
\docType{methods}
\name{ComputeErrorProb}
\alias{ComputeErrorProb}
\alias{ComputeErrorProb,contObsPredModelParam,ErrorProbParam-method}
\title{\code{ComputeErrorProb} computes the probability of measurement error for each statistical unit}
\usage{
ComputeErrorProb(object, Param)

\S4method{ComputeErrorProb}{contObsPredModelParam,ErrorProbParam}(object, Param)
}
\arguments{
\item{object}{Object of class \linkS4class{contObsPredModelParam} containing the statistical
units whose probability of measurement error for each variable is to be computed.}

\item{Param}{Object of virtual class \linkS4class{ErrorProbParam} with the parameters determining
the method of computation of the probability of measurement error of each statistical unit.}
}
\value{
Object of class \linkS4class{contObsPredModelParam} with the measurement error
probabilities computed for each variable and each statistical unit.
}
\description{
\code{ComputeErrorProb} computes the probability of measurement error for each statistical unit
}
\examples{
\dontrun{}
}

