% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeObsErrorSTD.R
\docType{methods}
\name{ComputeObsErrorSTD}
\alias{ComputeObsErrorSTD}
\alias{ComputeObsErrorSTD,contObsPredModelParam,ObsErrorSTDParam-method}
\title{\code{ComputeObsErrorSTD} computes the observation error variance for each statistical unit}
\usage{
ComputeObsErrorSTD(object, Param)

\S4method{ComputeObsErrorSTD}{contObsPredModelParam,ObsErrorSTDParam}(object,
  Param)
}
\arguments{
\item{object}{Object of class \linkS4class{contObsPredModelParam} containing the statistical
units whose probability of measurement error for each variable is to be computed.}

\item{Param}{Object of virtual class \linkS4class{ObsErrorSTDParam} with the parameters determining
the method of computation of the observation error variance of each statistical unit.}
}
\value{
Object of class \linkS4class{ObsErrorSTDParam} with the measurement error
probabilities computed for each variable and each statistical unit.
}
\description{
\code{ComputeObsErrorSTD} computes the observation error variance for each statistical unit
}
\examples{
\dontrun{

ObsPredPar <- new(Class = 'contObsPredModelParam',
                  Data = FD,
                  VarRoles = list(Units = 'NOrden', Domains = 'Tame_05._4.'))

ImpParam <- new(Class = 'MeanImputationParam',
                VarNames = c('CifraNeg_13.___', 'Personal_07.__2.__'),
                DomainNames =  c('Tame_05._4.'))
ObsErrVarMLEParam <- new(Class = 'ObsErrorSTDMLEParam',
                         RawData = FD.StQList,
                         EdData = FF.StQList,
                         VarNames = c('CifraNeg_13.___', 'Personal_07.__2.__'),
                         Imputation = ImpParam)
ObsPredPar <- ComputeObsErrorSTD(ObsPredPar, ObsErrVarMLEParam)

}
}
