% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeObsErrVar.R
\docType{methods}
\name{ComputeObsErrVar}
\alias{ComputeObsErrVar}
\alias{ComputeObsErrVar,contObsPredModelParam,ObsErrVarParam-method}
\title{\code{ComputeObsErrVar} computes the observation error variance for each statistical unit}
\usage{
ComputeObsErrVar(object, Param)

\S4method{ComputeObsErrVar}{contObsPredModelParam,ObsErrVarParam}(object, Param)
}
\arguments{
\item{object}{Object of class \linkS4class{contObsPredModelParam} containing the statistical
units whose probability of measurement error for each variable is to be computed.}

\item{Param}{Object of virtual class \linkS4class{ObsErrVarParam} with the parameters determining
the method of computation of the observation error variance of each statistical unit.}
}
\value{
Object of class \linkS4class{ObsErrVarParam} with the measurement error
probabilities computed for each variable and each statistical unit.
}
\description{
\code{ComputeObsErrVar} computes the observation error variance for each statistical unit
}
\examples{
\dontrun{

ObsPredPar <- new(Class = 'contObsPredModelParam',
                  Data = Data,
                  VarRoles = list(Units = 'NOrden', Domains = 'Tame_05._2.'))

ImpParam <- new(Class = 'MeanImputationParam',
                VarNames = c('CifraNeg_13.___', 'Personal_07.__2.__'),
                DomainNames =  c('Tame_05._2.', 'ActivEcono_35._4._2.1.4._0'))
ObsErrVarMLEParam <- new(Class = 'ObsErrVarMLEParam',
                         RawData = FF.StQList,
                         EdData = FD.StQList,
                         VarNames = c('CifraNeg_13.___', 'Personal_07.__2.__'),
                         Imputation = ImpParam)
ComputeErrorProb(ObsPredPar, ObsErrVarMLEParam)

}
}

