% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputePred.R
\docType{methods}
\name{ComputePred}
\alias{ComputePred}
\alias{ComputePred,contObsPredModelParam,PredParam-method}
\title{\code{ComputePred} computes the predicted values and the prediction error std for each
statistical unit}
\usage{
ComputePred(object, Param)

\S4method{ComputePred}{contObsPredModelParam,PredParam}(object, Param)
}
\arguments{
\item{object}{Object of class \linkS4class{contObsPredModelParam} containing the statistical
units whose predicted values and prediction error std for each variable is to be computed.}

\item{Param}{Object of virtual class \linkS4class{PredParam} with the parameters determining
the method of computation of the predicted values and prediction error std of each statistical
unit.}
}
\value{
Object of class \linkS4class{contObsPredModelParam} with the predicted values and
prediction error std computed for each variable and each statistical unit.
}
\description{
\code{ComputePred} computes the predicted values and the prediction error std for each
statistical unit
}
\examples{
\dontrun{
library(xlsx)
library(StQ)
library(RepoReadWrite)
DD <- RepoXLSToDD('S:/E30183/E30183.NombresVariables_V1.xlsx')
FD.StQ <- ReadRepoFile('S:/E30183/E30183.FD_V1.MM032016.P_1', DD, perl = TRUE)
FF.StQ <- ReadRepoFile('S:/E30183/E30183.FF_V1.MM022016.D_1', DD, perl = TRUE)
ObsPredPar <- new(Class = 'contObsPredModelParam',
                  Data = FD.StQ,
                  VarRoles = list(Units = 'NOrden', Domains = c('Tame_05._2.', 'ActivEcono_35._4._2.1.4._0')))
ImpParam <- new(Class = 'MeanImputationParam',
                VarNames = c('PredCifraNeg_13.___', 'PredErrorSTDCifraNeg_13.___',
                             'PredPersonal_07.__2.__', 'PredErrorSTDPersonal_07.__2.__'),
                DomainNames =  c('Tame_05._2.', 'ActivEcono_35._4._2.1.4._0'))
PredlmParam <- new(Class = 'PredlmParam',
                   EdData = FF.StQ,
                   VarNames = c('CifraNeg_13.___', 'Personal_07.__2.__'),
                   DomainNames = 'Tame_05._2.',
                   Imputation = ImpParam)


ComputePred(ObsPredPar, PredlmParam)

}
}

