% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\docType{methods}
\name{getData,contObsPredModelParam-method}
\alias{getData,contObsPredModelParam-method}
\title{Return slot \code{Data} from the input object}
\usage{
\S4method{getData}{contObsPredModelParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{contObsPredModelParam}.}
}
\value{
Object of class \linkS4class{StQ} corresponding to the slot \code{Data} of the input
parameter.
}
\description{
\code{getData} extracts the slot \code{Data} which contains the parameters of a
continuous observation-prediction model or all data necessary to compute them of the input object.
}
\examples{

}
