% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDomainNames.R
\docType{methods}
\name{getDomainNames}
\alias{getDomainNames}
\alias{getDomainNames,ObsErrorSTDParam-method}
\alias{getDomainNames,ErrorProbParam-method}
\alias{getDomainNames,PredValueTSParam-method}
\title{Return slot \code{DomainNames} from the input object.}
\usage{
getDomainNames(object)

\S4method{getDomainNames}{ObsErrorSTDParam}(object)

\S4method{getDomainNames}{ErrorProbParam}(object)

\S4method{getDomainNames}{PredValueTSParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{ObsErrorSTDParam}, \linkS4class{ErrorProbParam} or
\linkS4class{PredValueTSParam}.}
}
\value{
character vector corresponding to the slot \code{DomainNames} of the input parameter.
}
\description{
\code{getDomainNames} extracts the slot \code{DomainNames}, the names of the
variables determining the domains over which to compute the imputation method of missing values,
of the input object.

In the case of objects of classes \linkS4class{ObsErrorSTDParam} and \linkS4class{ErrorProbParam},
it returns the \code{DomainNames} of the slot \code{Imputation}.

In the case of objects of class \linkS4class{PredValueTSParam}, it returns the \code{DomainNames}
of the slot \code{ImputationParam}.
}
\examples{

}
