% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEdData.R
\docType{methods}
\name{getEdData}
\alias{getEdData}
\alias{getEdData,ObsErrorSTDParam-method}
\alias{getEdData,ErrorProbParam-method}
\title{Return slot \code{EdData} from the input object}
\usage{
getEdData(object)

\S4method{getEdData}{ObsErrorSTDParam}(object)

\S4method{getEdData}{ErrorProbParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{ObsErrorSTDParam} or \linkS4class{ErrorProbParam}.}
}
\value{
Object of class \linkS4class{StQList} corresponding to the slot \code{EdData} of the input
parameter.
}
\description{
\code{getEdData} extracts the slot \code{EdData} of the input object.
}
\examples{

}
