% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getErrorProb.R
\docType{methods}
\name{getErrorProb}
\alias{getErrorProb}
\alias{getErrorProb,contObsPredModelParam-method}
\title{Return component \code{ErroProb} from the input object}
\usage{
getErrorProb(object)

\S4method{getErrorProb}{contObsPredModelParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{ErrorProbParam}.}
}
\value{
character vector corresponding to the component \code{ErrorProb} of the input
parameter.
}
\description{
\code{getErroProb} extracts the component \code{ErrorProb} of the slot
\code{VarRoles} of the input object.
}
\examples{

}
