% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImputation.R
\docType{methods}
\name{getImputation}
\alias{getImputation}
\alias{getImputation,ObsErrorSTDParam-method}
\alias{getImputation,ErrorProbParam-method}
\title{Return slot \code{Imputation} from the input object}
\usage{
getImputation(object)

\S4method{getImputation}{ObsErrorSTDParam}(object)

\S4method{getImputation}{ErrorProbParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{ObsErrorSTDParam} or \linkS4class{ErrorProbParam}.}
}
\value{
Object of class \linkS4class{ImputationParam} corresponding to the slot with parameters
to impute of the input parameter.
}
\description{
\code{getImputation} extracts the slot \code{Imputation} of the input object, which
contains the parameters to impute missing values.
}
\examples{

}
