% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImputationParam.R
\docType{methods}
\name{getImputationParam}
\alias{getImputationParam}
\alias{getImputationParam,PredValueTSParam-method}
\title{Return slot \code{ImputationParam} from the input object}
\usage{
getImputationParam(object)

\S4method{getImputationParam}{PredValueTSParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{PredValueTSParam}.}
}
\value{
Object of class \linkS4class{ImputationParam} corresponding to the slot
\code{ImputationParam} of the input parameter.
}
\description{
\code{getImputationParam} extracts the slot \code{ImputationParam} of the input
object with the parameters for the imputation method of missing values.
}
\examples{

}
