% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getObjVariables.R
\docType{methods}
\name{getObjVariables}
\alias{getObjVariables}
\alias{getObjVariables,contObsPredModelParam-method}
\title{Return the component \code{ObjVariables} from the input object}
\usage{
getObjVariables(object)

\S4method{getObjVariables}{contObsPredModelParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{contObsPredModelParam}.}
}
\value{
character vector corresponding to the component \code{ObjVariables} of the input
parameter.
}
\description{
\code{getObjVariables} extracts the component \code{ObjVariables} of the slot
\code{VarRoles} of the input object.
}
\examples{

}
