% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getObsErrorSTD.R
\docType{methods}
\name{getObsErrorSTD}
\alias{getObsErrorSTD}
\alias{getObsErrorSTD,contObsPredModelParam-method}
\title{Return component \code{ObsErrSTD} from the input object}
\usage{
getObsErrorSTD(object)

\S4method{getObsErrorSTD}{contObsPredModelParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{contObsPredModelParam}.}
}
\value{
character vector corresponding to the component \code{ObsErrSTD} of the input parameter.
}
\description{
\code{getObsErrorSTD} extracts the component \code{ObsErrSTD} of the slot
\code{VarRoles} of the input object.
}
\examples{

}
