% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPredErrorSTD.R
\docType{methods}
\name{getPredErrorSTD}
\alias{getPredErrorSTD}
\alias{getPredErrorSTD,contObsPredModelParam-method}
\title{Return component \code{getPredErrorSTD} from the input object}
\usage{
getPredErrorSTD(object)

\S4method{getPredErrorSTD}{contObsPredModelParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{contObsPredModelParam}.}
}
\value{
character vector corresponding to the component \code{getPredErrorSTD} of the input
parameter.
}
\description{
\code{getPredErrorSTD} extracts the standard deviation of the predicted values
from the slot \code{VarRoles} of the input object.
}
\examples{

}
