% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPredictionParam.R
\docType{methods}
\name{getPredictionParam}
\alias{getPredictionParam}
\alias{getPredictionParam,PredValueTSParam-method}
\title{Return slot \code{PredictionParam} from the input object}
\usage{
getPredictionParam(object)

\S4method{getPredictionParam}{PredValueTSParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{PredValueTSParam}.}
}
\value{
slot of class \linkS4class{PredTSParam} corresponding to the slot \code{PrectionParam}
of the input parameter.
}
\description{
\code{getPredictionParam} extracts the the parameters to make predictions
\code{PredictionParam} of the input object.
}
\examples{

}
