% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPredValues.R
\docType{methods}
\name{getPredValues}
\alias{getPredValues}
\alias{getPredValues,contObsPredModelParam-method}
\title{Return component \code{PredValues} from the input object}
\usage{
getPredValues(object)

\S4method{getPredValues}{contObsPredModelParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{contObsPredModelParam}.}
}
\value{
character vector corresponding to the component \code{PredValues} of the input
parameter.
}
\description{
\code{getPredValues} extracts the predicted values from the slot \code{VarRoles}
of the input object.
}
\examples{

}
