% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRawData.R
\docType{methods}
\name{getRawData}
\alias{getRawData}
\alias{getRawData,ObsErrorSTDParam-method}
\alias{getRawData,ErrorProbParam-method}
\title{Return slot \code{RawData} from the input object}
\usage{
getRawData(object)

\S4method{getRawData}{ObsErrorSTDParam}(object)

\S4method{getRawData}{ErrorProbParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{ObsErrorSTDParam} or \linkS4class{ErrorProbParam}.}
}
\value{
Object of class \linkS4class{StQList} corresponding to the slot \code{RawData} of the input
parameter.
}
\description{
\code{getRawData} extracts the slot \code{RawData} of the input object.
}
\examples{

}
