% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarNames.R
\docType{methods}
\name{getVarNames}
\alias{getVarNames}
\alias{getVarNames,ObsErrorSTDParam-method}
\alias{getVarNames,ErrorProbParam-method}
\alias{getVarNames,PredValueTSParam-method}
\title{Return slot \code{VarNames} from the input object}
\usage{
getVarNames(object)

\S4method{getVarNames}{ObsErrorSTDParam}(object)

\S4method{getVarNames}{ErrorProbParam}(object)

\S4method{getVarNames}{PredValueTSParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{ObsErrorSTDParam}, \linkS4class{ErrorProbParam} or
\linkS4class{PredValueTSParam} .}
}
\value{
Character vector corresponding to the slot \code{VarNames} of the input parameter.
}
\description{
\code{getVarNames} extracts the slot \code{VarNames} of the input object.

In the case of objects of classes \linkS4class{ObsErrorSTDParam} and \linkS4class{ErrorProbParam},
it returns the slot \code{VarNames}.

In the case of objects of class \linkS4class{PredValueTSParam}, it returns the \code{VarNames}
of the slot \code{ImputationParam} which contains the names of the variables to be imputed.
}
\examples{

}
