% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarRoles.R
\docType{methods}
\name{getVarRoles}
\alias{getVarRoles}
\alias{getVarRoles,contObsPredModelParam-method}
\title{Return slot \code{VarRoles} from the input object}
\usage{
getVarRoles(object)

\S4method{getVarRoles}{contObsPredModelParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{contObsPredModelParam}.}
}
\value{
List with components \code{Units}, \code{Domains}, \code{DesignW},\code{ObjVariables},
\code{PredValues}, \code{PredErrorSTD}, \code{ObsErrorSTD},\code{ErrorProb}
}
\description{
\code{getVarRoles} extracts the slot \code{VarRoles} of the input object.
}
\examples{

}
