% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObsErrorSTDMLEParam-class.R
\docType{class}
\name{ObsErrorSTDMLEParam-class}
\alias{ObsErrorSTDMLEParam-class}
\title{S4 class for the parameters of the ML estimation of the obervation error variance}
\description{
Definition of the S4 class named \code{ObsErrorSTDMLEParam} for the parameters of
a continuous observation-prediction model in the optimization approach to selective editing.
}
\section{Slots}{

\describe{
\item{\code{RawData}}{\linkS4class{StQList} object with the raw data.}

\item{\code{EdData}}{\linkS4class{StQList} object with the edited data.}

\item{\code{VarNames}}{Character vector with the names of the variables whose probability errors are to
be computed.}

\item{\code{Imputation}}{\linkS4class{ImputationParam} object with the parameters to imputed missing
values during the computation of the error probabilities.}
}}
\examples{
# An empty ObsErrorSTDMLEParam object:
new(Class = 'ObsErrorSTDMLEParam')

\dontrun{
ImpParam <- new(Class = 'MeanImputationParam',
                VarNames = 'CifraNeg_13.___',
                DomainNames =  'Tame_05._4.')
ObsErrorSTDMLEParam <- new(Class = 'ObsErrorSTDMLEParam',
                         RawData = FD.StQList,
                         EdData = FG.StQList,
                         VarNames = 'CifraNeg_13.___',
                         Imputation = ImpParam)


}

}

