% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredValueTSParam-class.R
\docType{class}
\name{PredValueTSParam-class}
\alias{PredValueTSParam-class}
\title{S4 class for the parameters to compute the predicted values and the prediction error std
for each variable specified in the parameter for prediction}
\description{
Definition of the S4 class named \code{PredValueTSParam} for the parameters to
compute the predicted values and the prediction error std.
}
\section{Slots}{

\describe{
\item{\code{PredictionParam}}{\linkS4class{PredTSParam} object with the parameters to compute the
predicted values and their standard deviations.}

\item{\code{ImputationParam}}{\linkS4class{ImputationParam} object with the parameters to impute missing
values during the computation of the predicted values and their standard deviations.}
}}

\examples{
# An empty contObsPredModelParam object:
new(Class = 'PredValueTSParam')

}
