% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setData.R
\docType{methods}
\name{setData<-,contObsPredModelParam,StQ-method}
\alias{setData<-,contObsPredModelParam,StQ-method}
\title{Set value of slot \code{Data} of an object \linkS4class{contObsPredModelParam}}
\usage{
\S4method{setData}{contObsPredModelParam,StQ}(object) <- value
}
\arguments{
\item{object}{Object whose slot \code{Data} is to be assigned.}

\item{value}{\linkS4class{StQ} to be assigned to the slot \code{Data}.}
}
\value{
Object \linkS4class{contObsPredModelParam} with slot \code{Data} updated.
}
\description{
\code{setData} assigns a \linkS4class{StQ} to the slot \code{Data} of the input
object \linkS4class{contObsPredModelParam}
}
