% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDesignW.R
\docType{methods}
\name{setDesignW<-}
\alias{setDesignW<-}
\alias{setDesignW<-,contObsPredModelParam,data.table-method}
\title{\code{setDesignW} sets the design weight(s) for each variable and for each statistical unit}
\usage{
setDesignW(object) <- value

\S4method{setDesignW}{contObsPredModelParam,data.table}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{contObsPredModelParam} containing the statistical
units whose design weights are to be computed.}

\item{Param}{\linkS4class{data.table} with the design weigths for each variable and each
statistical unit.}
}
\value{
Object of class \linkS4class{contObsPredModelParam} with the design weights.
}
\description{
\code{setDesignW} sets the design weight(s) for each variable and for each statistical unit
}
\examples{
\dontrun{

load('C:/FDStQList.RData')
load('C:/FGStQList.RData')
FD <- FD.StQList[['MM112016']]
FDList <- FD.StQList[-11]
ObsPredPar <- new(Class = 'contObsPredModelParam',
                  Data = FD,
                  VarRoles = list(Units = 'NOrden', Domains = 'Tame_05._4.'))
Weights <- StQ::dcast_StQ(ObsPredPar@Data)[, 'NOrden', with = FALSE][, Personal_07._2.___ := 1][, CifraNeg_13.___ := 1]
setDesignW(ObsPredPar) <- Weights

}
}

