% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setErrorProb.R
\name{setErrorProb<-}
\alias{setErrorProb<-}
\title{Set value of component \code{ErrorProb} of an object \linkS4class{contObsPredModelParam}}
\usage{
setErrorProb(object) <- value
}
\arguments{
\item{object}{Object \linkS4class{contObsPredModelParam} whose component \code{ErrorProb}
is to be assigned.}

\item{value}{character vector to be assigned to the component \code{ErrorProb}.}
}
\value{
Object \linkS4class{contObsPredModelParam} with the component \code{ErrorProb} updated.
}
\description{
\code{setErrorProb} assigns a character value to the component \code{ErrorProb} of
the slot \code{VarRoles} of the input object \linkS4class{contObsPredModelParam}.
}
