% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setObjVariables.R
\name{setObjVariables<-}
\alias{setObjVariables<-}
\title{Set value of component \code{ObjVariables} of an object \linkS4class{contObsPredModelParam}}
\usage{
setObjVariables(object) <- value
}
\arguments{
\item{object}{Object \linkS4class{contObsPredModelParam} whose component \code{ObjVariables}
is to be assigned.}

\item{value}{character vector to be assigned to the component \code{ObjVariables}.}
}
\value{
Object \linkS4class{contObsPredModelParam} with the component \code{ObjVariables} updated.
}
\description{
\code{setObjVariables} assigns a character value to the component \code{ObjVariables}
of the slot \code{VarRoles} of the input object \linkS4class{contObsPredModelParam}.
}
