% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setobsErrorSTD.R
\name{setObsErrorSTD<-}
\alias{setObsErrorSTD<-}
\title{Set value of component \code{ObsErrSTD} of an object \linkS4class{contObsPredModelParam}}
\usage{
setObsErrorSTD(object) <- value
}
\arguments{
\item{object}{Object \linkS4class{contObsPredModelParam} whose component \code{ObsErrSTD} is to
be assigned.}

\item{value}{character vector to be assigned to the component \code{ObsErrSTD}.}
}
\value{
Object \linkS4class{contObsPredModelParam} with the component \code{ObsErrSTD} updated.
}
\description{
\code{setObsErrorSTD} assigns a character value to the component \code{ObsErrSTD}
of the slot \code{VarRoles} of the input object \linkS4class{contObsPredModelParam}.
}
