% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPredValues.R
\name{setPredValues<-}
\alias{setPredValues<-}
\title{Set value of component \code{PredValues} of an object \linkS4class{contObsPredModelParam}}
\usage{
setPredValues(object) <- value
}
\arguments{
\item{object}{Object \linkS4class{contObsPredModelParam} whose component \code{PredValues}
is to be assigned.}

\item{value}{character vector to be assigned to the component \code{PredValues}.}
}
\value{
Object \linkS4class{contObsPredModelParam} with the component \code{PredValues} updated.
}
\description{
\code{setPredValues} assigns a character value to the component \code{PredValues} of
the slot \code{VarRoles} of the input object \linkS4class{contObsPredModelParam}.
}
