% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contObsPredModelParam-class.R
\docType{class}
\name{contObsPredModelParam-class}
\alias{contObsPredModelParam-class}
\title{S4 class for the parameters of a continuous observation-prediction model}
\description{
Definition of the S4 class named \code{contObsPredModelParam} for the parameters of
a continuous observation-prediction model in the optimization approach to selective editing.
}
\section{Slots}{

\describe{
\item{\code{Data}}{\linkS4class{data.table} with the parameters or all data necessary to compute them.}

\item{\code{VarRoles}}{List with components \code{Units}, \code{Domains}, \code{DesignW},
\code{ObjVariables}, \code{PredValues}, \code{PredErrorSTD}, \code{ObsErrorSTD},
\code{ErrorProb} being character vectors containing the column names according to their
respective role in the model.}
}}

\examples{
# An empty contObsPredModelParam object:
new(Class = 'contObsPredModelParam')

\dontrun{

ObsPredPar <- new(Class = 'contObsPredModelParam',
                  Data = FD,
                  VarRoles = list(Units = 'NOrden', Domains = 'Tame_05._2.'))

}

}
