% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ObsPredPar.R
\docType{methods}
\name{plot,contObsPredModelParam,character-method}
\alias{plot,contObsPredModelParam,character-method}
\title{Plot the main values in the optimization approach to selective editing for the specified
units.}
\usage{
\S4method{plot}{contObsPredModelParam,character}(x, y, ...)
}
\arguments{
\item{x}{Object of class \linkS4class{contObsPredModelParam}.}

\item{y}{Charactec vector with the names of the variable names.}

\item{...}{Arguments to be passed to methods, especially including

\itemize{

 \item \code{Query} as a \linkS4class{data.table} with the values of the identification
 variables of the queried statistical units;

 \item \code{FT} as a \linkS4class{StQ} object with the values of the selected units in the
 optimization approach to selective editing;

 \item \code{ErrorMoments} as a \code{numeric} vector with the error moments corresponding to
 queried units;
}}
}
\value{
Returns an invisible \code{NULL} plotting the one scatterplot per variable.
}
\description{
This method overloads the plot function to draw a picture with the information about
the queried units in the optimization approach to selective editing.
}
\examples{

}
